+++ app/wyzecam/tutk/tutk.py
@@ -534,6 +534,7 @@ def av_recv_frame_data(
     int,
     Optional[bytes],
     Optional[Union[FrameInfoStruct, FrameInfo3Struct]],
+    Optional[int],
 ]:
     """A new version AV client receives frame data from an AV server.
 
@@ -566,7 +567,7 @@ def av_recv_frame_data(
     )
 
     if errno < 0:
-        return errno, None, None
+        return errno, None, None, None
 
     frame_data_actual: bytes = frame_data[: frame_data_actual_len.value]  # type: ignore
     frame_info_actual: Union[FrameInfoStruct, FrameInfo3Struct]
@@ -579,7 +580,7 @@ def av_recv_frame_data(
             f"Unknown frame info structure format! len={frame_info_actual_len}"
         )
 
-    return (0, frame_data_actual, frame_info_actual)
+    return (0, frame_data_actual, frame_info_actual, frame_index.value)
 
 
 def av_recv_io_ctrl(
@@ -672,6 +673,17 @@ def av_client_clean_local_video_buf(tutk_platform_lib: CDLL, channel_id: c_int) 
     :param channel_id: The channel ID of the AV channel to clean buffer
     """
     tutk_platform_lib.avClientCleanLocalVideoBuf(channel_id)
+
+
+def av_client_clean_local_audio_buf(tutk_platform_lib: CDLL, channel_id: c_int) -> None:
+    """Clean the local audio buffer of the client.
+
+    This function is used to clean the audio buffer that the client
+    has already received
+
+    :param channel_id: The channel ID of the AV channel to clean buffer
+    """
+    tutk_platform_lib.avClientCleanAudioBuf(channel_id)
 
 
 def av_client_stop(tutk_platform_lib: CDLL, av_chan_id: c_int) -> None:
@@ -885,7 +897,11 @@ def iotc_connect_by_uid_parallel(
 
 
 def iotc_connect_by_uid_ex(
-    tutk_platform_lib: CDLL, p2p_id: str, session_id: c_int, auth_key: bytes
+    tutk_platform_lib: CDLL,
+    p2p_id: str,
+    session_id: c_int,
+    auth_key: bytes,
+    timeout: int = 20,
 ) -> c_int:
     """Used by a client to connect a device.
 
@@ -903,6 +919,7 @@ def iotc_connect_by_uid_ex(
     connect_input = St_IOTCConnectInput()
     connect_input.cb = sizeof(connect_input)
     connect_input.auth_key = auth_key
+    connect_input.timeout = timeout
 
     resultant_session_id: c_int = tutk_platform_lib.IOTC_Connect_ByUIDEx(
         c_char_p(p2p_id.encode("ascii")), session_id, byref(connect_input)